fix bundling issues (#34387)
fixes: #34386
