+++ js/apps/admin-ui/src/components/routable-tabs/RoutableTabs.tsx
@@ -22,7 +22,7 @@ import {
 } from "react-router-dom";
 import { useServerInfo } from "../../context/server-info/ServerInfoProvider";
 import { PageHandler } from "../../page/PageHandler";
-import { TAB_PROVIDER } from "../../page/PageList";
+import { TAB_PROVIDER } from "../../page/constants";
 import useIsFeatureEnabled, { Feature } from "../../utils/useIsFeatureEnabled";
 import { useTranslation } from "react-i18next";
