+++ js/apps/admin-ui/src/page/constants.ts
@@ -1,0 +1,2 @@
+export const PAGE_PROVIDER = "org.keycloak.services.ui.extend.UiPageProvider";
+export const TAB_PROVIDER = "org.keycloak.services.ui.extend.UiTabProvider";
