+++ js/apps/admin-ui/vite.config.ts
@@ -25,6 +25,7 @@ export default defineConfig(({ mode }) => {
       }
     : {
         outDir: "target/classes/theme/keycloak.v2/admin/resources",
+        external: ["src/index.ts"],
       };
   return {
     base: "",
